+++ cedar-policy-validator/src/rbac.rs
@@ -402,7 +402,7 @@ mod test {
             Annotations, Effect, Eid, EntityUID, Expr, PolicyID, PrincipalConstraint,
             ResourceConstraint,
         },
-        parser::{parse_policy, parse_policy_template},
+        parser::{parse_policy, parse_policy_or_template},
         test_utils::{expect_err, ExpectedErrorMessageBuilder},
     };
     use miette::Report;
@@ -417,7 +417,7 @@ mod test {
     #[test]
     fn validate_entity_type_empty_schema() {
         let src = r#"permit(principal, action, resource == foo_type::"foo_name");"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
         let validate = Validator::new(ValidatorSchema::empty());
         let notes: Vec<ValidationError> = validate.validate_entity_types(&policy).collect();
         expect_err(
@@ -460,7 +460,7 @@ mod test {
         let schema = schema_file.try_into().unwrap();
 
         let src = r#"permit(principal == admins::"admin1", action == Action::"act", resource == bin::"bin");"#;
-        let p = parse_policy_template(None, src).unwrap();
+        let p = parse_policy_or_template(None, src).unwrap();
 
         let validate = Validator::new(schema);
         let notes: Vec<ValidationError> =
@@ -529,7 +529,7 @@ mod test {
         let singleton_schema = schema_file.try_into().unwrap();
 
         let src = r#"permit(principal, action, resource == bar_type::"bar_name");"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
         let validate = Validator::new(singleton_schema);
         let notes: Vec<ValidationError> = validate.validate_entity_types(&policy).collect();
         expect_err(
@@ -548,7 +548,7 @@ mod test {
     #[test]
     fn validate_action_id_empty_schema() {
         let src = r#"permit(principal, action == Action::"foo_name", resource);"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
         let validate = Validator::new(ValidatorSchema::empty());
         let notes: Vec<ValidationError> = validate.validate_action_ids(&policy).collect();
         expect_err(
@@ -706,7 +706,7 @@ mod test {
         let singleton_schema = schema_file.try_into().unwrap();
 
         let src = r#"permit(principal, action == Action::"bar_name", resource);"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
         let validate = Validator::new(singleton_schema);
         let notes: Vec<ValidationError> = validate.validate_action_ids(&policy).collect();
         expect_err(
@@ -769,7 +769,7 @@ mod test {
         let schema = descriptors.try_into().unwrap();
 
         let src = r#"permit(principal, action == Bogus::Action::"foo_name", resource);"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
         let validate = Validator::new(schema);
         let notes: Vec<ValidationError> = validate.validate_action_ids(&policy).collect();
         expect_err(
@@ -835,7 +835,7 @@ mod test {
         let schema = descriptors.try_into().unwrap();
 
         let src = r#"permit(principal == Bogus::Foo::"bar", action, resource);"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
         let validate = Validator::new(schema);
         let notes: Vec<ValidationError> = validate.validate_entity_types(&policy).collect();
         expect_err(
@@ -1082,7 +1082,7 @@ mod test {
 
         let src =
             r#"permit(principal == baz::"p", action == Action::"foo", resource == baz::"r");"#;
-        let p = parse_policy_template(None, src).unwrap();
+        let p = parse_policy_or_template(None, src).unwrap();
 
         let validate = Validator::new(schema);
         let notes: Vec<ValidationError> =
@@ -1106,7 +1106,7 @@ mod test {
 
         let src =
             r#"permit(principal == bar::"p", action == Action::"foo", resource == bar::"r");"#;
-        let p = parse_policy_template(None, src).unwrap();
+        let p = parse_policy_or_template(None, src).unwrap();
 
         let validate = Validator::new(schema);
         let notes: Vec<ValidationError> =
@@ -1130,7 +1130,7 @@ mod test {
 
         let src =
             r#"permit(principal == baz::"p", action == Action::"foo", resource == bar::"r");"#;
-        let p = parse_policy_template(None, src).unwrap();
+        let p = parse_policy_or_template(None, src).unwrap();
 
         let validate = Validator::new(schema);
         let notes: Vec<ValidationError> =
@@ -1153,12 +1153,12 @@ mod test {
         let (_, _, _, schema) = schema_with_single_principal_action_resource();
 
         let policy =
-            parse_policy_template(None, "permit(principal is bar, action, resource);").unwrap();
+            parse_policy_or_template(None, "permit(principal is bar, action, resource);").unwrap();
 
         let validator = Validator::new(schema);
         assert_validate_policy_succeeds(&validator, &policy);
 
-        let policy = parse_policy_template(
+        let policy = parse_policy_or_template(
             None,
             r#"permit(principal is bar in bar::"baz", action, resource);"#,
         )
@@ -1172,7 +1172,7 @@ mod test {
         let (_, _, _, schema) = schema_with_single_principal_action_resource();
 
         let src = "permit(principal is baz, action, resource);";
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
 
         let validator = Validator::new(schema);
         assert_validate_policy_fails(
@@ -1188,7 +1188,7 @@ mod test {
         assert_validate_policy_flags_impossible_policy(&validator, &policy);
 
         let src = r#"permit(principal is biz in faz::"a", action, resource);"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
 
         assert_validate_policy_fails(
             &validator,
@@ -1217,7 +1217,7 @@ mod test {
         assert_validate_policy_flags_impossible_policy(&validator, &policy);
 
         let src = r#"permit(principal is bar in baz::"buz", action, resource);"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
 
         assert_validate_policy_fails(
             &validator,
@@ -1237,12 +1237,12 @@ mod test {
         let (_, _, _, schema) = schema_with_single_principal_action_resource();
 
         let policy =
-            parse_policy_template(None, "permit(principal, action, resource is baz);").unwrap();
+            parse_policy_or_template(None, "permit(principal, action, resource is baz);").unwrap();
 
         let validator = Validator::new(schema);
         assert_validate_policy_succeeds(&validator, &policy);
 
-        let policy = parse_policy_template(
+        let policy = parse_policy_or_template(
             None,
             r#"permit(principal, action, resource is baz in baz::"bar");"#,
         )
@@ -1256,7 +1256,7 @@ mod test {
         let (_, _, _, schema) = schema_with_single_principal_action_resource();
 
         let src = "permit(principal, action, resource is bar);";
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
 
         let validator = Validator::new(schema);
         assert_validate_policy_fails(
@@ -1272,7 +1272,7 @@ mod test {
         assert_validate_policy_flags_impossible_policy(&validator, &policy);
 
         let src = r#"permit(principal, action, resource is baz in bar::"buz");"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
 
         assert_validate_policy_fails(
             &validator,
@@ -1287,7 +1287,7 @@ mod test {
         assert_validate_policy_flags_impossible_policy(&validator, &policy);
 
         let src = r#"permit(principal, action, resource is biz in faz::"a");"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
 
         assert_validate_policy_fails(
             &validator,
@@ -1320,7 +1320,7 @@ mod test {
     fn is_unknown_entity_condition() {
         let (_, _, _, schema) = schema_with_single_principal_action_resource();
         let src = r#"permit(principal, action, resource) when { resource is biz };"#;
-        let policy = parse_policy_template(None, src).unwrap();
+        let policy = parse_policy_or_template(None, src).unwrap();
 
         let validator = Validator::new(schema);
         let err = validator