+++ cedar-policy-validator/src/typecheck/test/policy.rs
@@ -23,7 +23,7 @@ use std::sync::Arc;
 
 use cedar_policy_core::{
     ast::{EntityUID, Expr, PolicyID, Template, Var},
-    parser::{parse_policy, parse_policy_template},
+    parser::{parse_policy, parse_policy_or_template},
 };
 use smol_str::SmolStr;
 
@@ -177,7 +177,7 @@ fn entity_literal_typechecks() {
 
 #[test]
 fn policy_checked_in_multiple_envs() {
-    let t = parse_policy_template(
+    let t = parse_policy_or_template(
         Some(PolicyID::from_string("0")),
         r#"permit(principal, action == Action::"view_photo", resource) when { resource.file_type == "jpg" };"#
     ).expect("Policy should parse.");
@@ -205,7 +205,7 @@ fn policy_checked_in_multiple_envs() {
             == 1
     );
 
-    let t = parse_policy_template(
+    let t = parse_policy_or_template(
         Some(PolicyID::from_string("0")),
         r#"permit(principal, action == Action::"delete_group", resource) when { resource.file_type == "jpg" };"#
     ).expect("Policy should parse.");
@@ -1085,7 +1085,7 @@ mod templates {
     #[test]
     fn principal_eq_slot() {
         assert_policy_typechecks_simple_schema(
-            parse_policy_template(
+            parse_policy_or_template(
                 None,
                 r#"permit(principal == ?principal, action, resource);"#,
             )
@@ -1096,7 +1096,7 @@ mod templates {
     #[test]
     fn resource_eq_slot() {
         assert_policy_typechecks_simple_schema(
-            parse_policy_template(None, r#"permit(principal, action, resource == ?resource);"#)
+            parse_policy_or_template(None, r#"permit(principal, action, resource == ?resource);"#)
                 .unwrap(),
         );
     }
@@ -1104,7 +1104,7 @@ mod templates {
     #[test]
     fn principal_resource_eq_slot() {
         assert_policy_typechecks_simple_schema(
-            parse_policy_template(
+            parse_policy_or_template(
                 None,
                 r#"permit(principal == ?principal, action, resource == ?resource);"#,
             )
@@ -1115,7 +1115,7 @@ mod templates {
     #[test]
     fn principal_in_slot() {
         assert_policy_typechecks_simple_schema(
-            parse_policy_template(
+            parse_policy_or_template(
                 None,
                 r#"permit(principal in ?principal, action, resource);"#,
             )
@@ -1126,7 +1126,7 @@ mod templates {
     #[test]
     fn resource_in_slot() {
         assert_policy_typechecks_simple_schema(
-            parse_policy_template(None, r#"permit(principal, action, resource in ?resource);"#)
+            parse_policy_or_template(None, r#"permit(principal, action, resource in ?resource);"#)
                 .unwrap(),
         );
     }
@@ -1134,7 +1134,7 @@ mod templates {
     #[test]
     fn principal_resource_in_slot() {
         assert_policy_typechecks_simple_schema(
-            parse_policy_template(
+            parse_policy_or_template(
                 None,
                 r#"permit(principal in ?principal, action, resource in ?resource);"#,
             )
@@ -1145,7 +1145,7 @@ mod templates {
     #[test]
     fn resource_slot_safe_body() {
         assert_policy_typechecks_simple_schema(
-            parse_policy_template(
+            parse_policy_or_template(
                 None,
                 r#"permit(principal, action, resource in ?resource) when { resource in Group::"Friends" && resource.name like "*" };"#,
             )
@@ -1156,7 +1156,7 @@ mod templates {
     #[test]
     fn resource_slot_error_body() {
         assert_policy_typecheck_fails_simple_schema(
-            parse_policy_template(
+            parse_policy_or_template(
                 None,
                 r#"permit(principal, action, resource in ?resource) when { resource in Group::"Friends" && resource.bogus };"#,
             )
@@ -1174,7 +1174,7 @@ mod templates {
     #[test]
     fn principal_slot_safe_body() {
         assert_policy_typechecks_simple_schema(
-            parse_policy_template(
+            parse_policy_or_template(
                 None,
                 r#"permit(principal == ?principal, action, resource in ?resource) when { principal has age && principal.age > 0};"#,
             )
@@ -1185,7 +1185,7 @@ mod templates {
     #[test]
     fn principal_slot_error_body() {
         assert_policy_typecheck_fails_simple_schema(
-            parse_policy_template(
+            parse_policy_or_template(
                 None,
                 r#"permit(principal == ?principal, action, resource) when { principal has age && principal.bogus > 0 };"#,
             )
@@ -1202,7 +1202,7 @@ mod templates {
 
     #[test]
     fn template_all_false() {
-        let template = parse_policy_template(
+        let template = parse_policy_or_template(
             None,
             r#"permit(principal == ?principal, action, resource) when { false };"#,
         )